Fixed backprop bug
bug was in bin_matrix_mult_fpvec (binarised_bp_test.c).

This multiplication was being done using the output vector incrementer in the outer loop, and the input in the inner loop.

e.g. if the network has a structure of 5, 5, then 2 layers and this was the first round of backprop then this function tries to multiply the activations of the last layer (shape: (1, 2)) by the binary weights of that layer (shape: (2, 5)). Binary eights, however, occur in rows, so it to calculate an entire output node, then move onto the next, you would have to access multiple weight rows, then the same multiple weight rows again when you calculate the next output node. Instead I'm calculating according to one input node and therefore one row at a time.