AXI redone - simulation shows inputs being written to weight ram but still has issues
The weight count variable goes up twice as fast as it's supposed to, leading to issues. Will look into tomorrow.
But the good news is that the synthesis recognises all the correct parts and isn't ignoring any counters that we need and is able to do some simulation without crashing :)